don't ask why, but data close to INT_MIN was overflowing, so I change it to longÈ

+++ includes/ps.h
@@ -6,7 +6,7 @@
 /*   By: adelille </var/mail/adelille>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/19 14:58:11 by adelille          #+#    #+#             */
-/*   Updated: 2021/10/09 19:27:17 by adelille         ###   ########.fr       */
+/*   Updated: 2021/10/09 19:45:30 by adelille         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 typedef struct s_stack
 {
-	int				data;
+	long			data;
 	int				index;
 	int				keep;
 	struct s_stack	*next;
